+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed a bug where Webhook Signing Secret setting could not be set to an environment variable. ([#61](https://github.com/craftcms/commerce-stripe/issues/61))
+- Fixed a bug where Commerce wouldn't correctly refund purchases by guest users using Payment Intents gateway. ([#64](https://github.com/craftcms/commerce-stripe/issues/64))
 
 ## 2.0.1.2 - 2019-07-08