+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Stripe for Craft Commerce
 
-## Unreleased
+## 2.1.0 - 2019-07-24
 
 ### Changed
 - Update Craft Commerce requirements to allow for Craft Commerce 3.